Adds get_connection_string tool to MCP server (#24)
We can have a tool that simply returns the connection string. The
problem with connection strings is that they are much more complex than
they appear on surface. You have to pick:
- Project ID
- Branch
- Compute
- Default Database
- Role
So, in theory, a proper tool to get a connection string would need to
receive these 5 thing as input, with ALL of them being optional because
there are cases when there is only one available choice for ALL of them,
in which case the MCP server should just be able to return the
connection string for:
- Your only project
- Its only branch
- Its only endpoint/compute
- Its only database
- The only role (probably neondb_owner)

I decided to make Project ID mandatory, I'm not sure having that as
optional is a good idea.

CC @Shridhad