Adds better hint to the list projects tool (#50)
I tried to make a test for this, but unfortunately it's hard to verify
that this is _actually_ working as a fix. Most of the time, Claude is
smart and it will pass the `search` parameter which means that the limit
is not actually a problem.

So, we'll have to vibe-trust this PR unfortunately.

CC @Shridhad

---------

Co-authored-by: Shridhar Deshmukh <shridhar@neon.tech>